requirements.txt
google-genai
python-dotenv
scenedetect[opencv]
opencv-python-headless
openai-whisper
numpy
Pillow
